+++ README.md
@@ -127,7 +127,7 @@
 |   | [ACM ICPC Team](https://www.hackerrank.com/challenges/acm-icpc-team)| [Java](https://github.com/RyanFehr/HackerRank/blob/master/Algorithms/Implementation/ACM%20ICPC%20Team/Solution.java) | _O(n^3)_ | _O(n)_ | Easy | 25| ||
 |   | [Taum and B'day](https://www.hackerrank.com/challenges/taum-and-bday)| [Java](https://github.com/RyanFehr/HackerRank/blob/master/Algorithms/Implementation/Taum%20and%20B'day/Solution.java) | _O(1)_ | _O(1)_ | Easy | 25| ||
 |   | [Organizing Containers of Balls](https://www.hackerrank.com/challenges/organizing-containers-of-balls)| [Java](https://github.com/RyanFehr/HackerRank/blob/master/Algorithms/Implementation/Organizing%20Containers%20of%20Balls/Solution.java) | _O(n^2)_ | _O(n^2)_ | Medium | 30| ||
-|   | [Enryption](https://www.hackerrank.com/challenges/encryption)| [Java](https://github.com/RyanFehr/HackerRank/blob/master/Algorithms/Implementation/Encyrption/Solution.java) | _O(n)_ | _O(n)_ | Medium | 30| ||
+|   | [Enryption](https://www.hackerrank.com/challenges/encryption)| [Java](https://github.com/RyanFehr/HackerRank/blob/master/Algorithms/Implementation/Encryption/Solution.java) | _O(n)_ | _O(n)_ | Medium | 30| ||
 |   | [Bigger is Greater](https://www.hackerrank.com/challenges/bigger-is-greater)| [Java](https://github.com/RyanFehr/HackerRank/blob/master/Algorithms/Implementation/Bigger%20is%20Greater/Solution.java) | _O(n)_ | _O(n)_ | Medium | 35| ||
 |   | [Modified Kaprekar Numbers](https://www.hackerrank.com/challenges/kaprekar-numbers)| [Java](https://github.com/RyanFehr/HackerRank/blob/master/Algorithms/Implementation/Modified%20Kaprekar%20Numbers/Solution.java) | _O(n)_ | _O(1)_ | Easy | 30| ||
 |   | [Beautiful Triplets](https://www.hackerrank.com/challenges/beautiful-triplets)| [Java](https://github.com/RyanFehr/HackerRank/blob/master/Algorithms/Implementation/Beautiful%20Triplets/Solution.java) | _O(n)_ | _O(n)_ | Easy | 20| ||
